Refactor transitions and update animation variables
