+++ src/components/Transitions/index.js
@@ -1,2 +1,5 @@
 export { default as Fade } from './Fade';
 export { default as Wipe } from './Wipe';
+export { default as Appear } from './Appear';
+export { default as Folder } from './Folder';
+export { default as Screen } from './Screen';
