+++ src/components/index.js
@@ -1,3 +1,2 @@
-export { default as ScreenTransition } from './ScreenTransition';
 export { default as Timeline } from './Timeline';
 export { default as Rules } from './Rules';
