+++ src/styles/components/guided/_section.scss
@@ -2,7 +2,8 @@
   flex-direction: row;
   width: 100%;
   padding: 0 35% 0 0;
-  transition: opacity var(--animation-standard-duration) var(--animation-default-easing), transform var(--animation-standard-duration) var(--animation-default-easing);
+  transition: opacity var(--animation-duration-standard-ms) var(--animation-easing),
+    transform var(--animation-duration-standard-ms) var(--animation-easing);
 
   &:first-child {
     padding-top: unit(4);
