+++ src/styles/components/rule/_rule.scss
@@ -107,8 +107,8 @@
     &--attribute {
       background: palette('rule-attribute');
       max-width: 500px;
-      transition: max-width var(--animation-default-easing) var(--animation-standard-duration),
-        padding var(--animation-default-easing) var(--animation-standard-duration);
+      transition: max-width var(--animation-easing) var(--animation-duration-standard-ms),
+        padding var(--animation-easing) var(--animation-duration-standard-ms);
 
       &:empty {
         max-width: 0;
