+++ src/styles/containers/_protocol.scss
@@ -15,7 +15,7 @@
     padding-bottom: 10px;
     background-color: palette('background');
     transform: translateZ(0) translateY(100%);
-    transition: transform var(--animation-default-easing) var(--animation-fast-duration);
+    transition: transform var(--animation-easing) var(--animation-duration-fast-ms);
   }
 
   &--has-changes {
